Continued to work on the voki taking service update

+++ backend/VokiTakingServicesCollection/GeneralVokiTakingService/src/GeneralVokiTakingService.Domain/voki_taking_session_aggregate/BaseVokiTakingSession.cs
@@ -43,7 +43,7 @@ public abstract class BaseVokiTakingSession : AggregateRoot<VokiTakingSessionId>
             (QuestionOrderInVokiTakingSession Order, ImmutableHashSet<GeneralVokiAnswerId> SavedAnsweres)
         > QuestionsToShow
     );
-    public abstract int QuestionsWithSavedAnswersCount();
+    public abstract ushort QuestionsWithSavedAnswersCount();
 
     public abstract VokiTakingStateToContinueFromSaved GetSavedStateToContinueTaking();
